refactor(usecase): rename file to keep defaults
