+++ src/core/usecase/GenerateKinesisEvents/GenerateKinesisEventsOutput.ts
@@ -1,0 +1,8 @@
+import { GenerateKinesisEventsInput } from '.';
+
+export interface GenerateKinesisEventsOutput extends GenerateKinesisEventsInput {
+  /**
+   * Total of loaded JSON to kinesis stream
+   */
+  loadedRecords: number;
+}
